refactor(challenge): 나열형 거대 파일 4개 컴포넌트별 분리 (Phase 3)
상태·쿼리 없는 프레젠테이션 컴포넌트 나열형 4개를 파일명 기반 하위폴더로 분리:
- Portfolio다른프로그램추천 728→495줄 (other-program-recommend/)
- ChallengePointView 664→431줄 (challenge-point-view/)
- ChallengeCheckList 583→245줄 (challenge-check-list/)
- ChallengeResult 399→265줄 (challenge-result/)
순수 이동/추출로 동작 변경 없음. 외부 importer 없음(내부 전용 컴포넌트).
