+++ apps/web/src/domain/program/challenge/challenge-view/ChallengeResult.tsx
@@ -1,109 +1,22 @@
 import { useMediaQuery } from '@mui/material';
-import { CSSProperties, ReactNode, useMemo } from 'react';
+import { useMemo } from 'react';
 import { FaCheck } from 'react-icons/fa6';
 
 import { Break } from '@/common/Break';
 import Heading2 from '@/common/header/Heading2';
 import { challengeColors } from '@/domain/program/challenge/ChallengeView';
-import Box from '@/domain/program/program-detail/Box';
 import SuperTitle from '@/domain/program/program-detail/SuperTitle';
-import { twMerge } from '@/lib/twMerge';
 import { ChallengeType, challengeTypeSchema } from '@/schema';
-
-const superTitle = '이 모든 고민을 한번에 해결!';
-
-const RESUME_CAREER_START_CONTENT = [
-  {
-    beforeImg: '/images/resume-start-before1.svg',
-    beforeCaption: '정리되지 않은 날 것의 경험과 역량들',
-    afterImg: '/images/resume-start-after1.svg',
-    afterCaption: '구체적이고 직무 역량이 구조화된 경험',
-  },
-  {
-    beforeImg: '/images/resume-start-before2.svg',
-    beforeCaption: '경험이 나열되기만 한 이력서',
-    afterImg: '/images/resume-start-after2.svg',
-    afterCaption: '업무의 세부 사항부터 역량까지 돋보이는 이력서',
-  },
-  {
-    beforeImg: '/images/resume-start-before3.svg',
-    beforeCaption: '역량이 보이지 않는 낡은 이력서 양식',
-    afterImg: '/images/resume-start-after3.svg',
-    afterCaption: '역량이 돋보이는 이력서 템플릿 활용',
-  },
-];
-
-const PERSONAL_STATEMENT_CONTENT = [
-  {
-    beforeImg: '/images/personal-statement-before1.jpg',
-    beforeCaption: '누구나 쓸 수 있는 추상적인 지원동기',
-    afterImg: '/images/personal-statement-after1.jpg',
-    afterCaption: 'A사의 가치 기술과 관련된 경험 연결',
-  },
-  {
-    beforeImg: '/images/personal-statement-before2.jpg',
-    beforeCaption: '직무 경험을 구구절절 나열하는 방식',
-    afterImg: '/images/personal-statement-after2.jpg',
-    afterCaption: '직무 키워드 선정 후, 관련된 경험 구체화',
-  },
-];
-
-const PORTFOLIO_CONTENT = [
-  {
-    beforeImg: '/images/up-1-before-494-302.png',
-    beforeCaption: '마치 서비스 소개서처럼 솔루션만 설명',
-    afterImg: '/images/up-1-after-494-302.png',
-    afterCaption: '유저 인터뷰 등을 통한 UX/UI 변경 부분 작성',
-  },
-  {
-    beforeImg: '/images/up-2-before-494-302.png',
-    beforeCaption: '어떤 활동을 했는지 경험만 나열',
-    afterImg: '/images/up-2-after-494-302.png',
-    afterCaption: '문제점 → 전략 → 솔루션을 보여주는 구조화',
-  },
-];
-
-const CAREER_START_CONTENT = [
-  {
-    beforeImg: '/images/career-start-before1.jpg',
-    beforeCaption: '매번 마감 기한 전까지 소재 고민하기',
-    afterImg: '/images/career-start-after1.jpg',
-    afterCaption: '경험 서랍에서 바로 직무 FIT한 소재 뽑기',
-  },
-  {
-    beforeImg: '/images/career-start-before2.jpg',
-    beforeCaption: '주변 말만 듣고 직무 고르기',
-    afterImg: '/images/career-start-after2.jpg',
-    afterCaption: '업무부터 역량까지 직무의 A to Z를 직접 찾고 결정하기',
-  },
-  {
-    beforeImg: '/images/career-start-before3.jpg',
-    beforeCaption: '양으로 승부 보는 이력서',
-    afterImg: '/images/career-start-after3.jpg',
-    afterCaption: '핵심 역량과 매력만 깔끔하게 간추린 이력서',
-  },
-];
-
-const EXPERIENCE_SUMMARY_CONTENT = [
-  {
-    beforeImg: '/images/experience-summary-before1.jpg',
-    beforeCaption: '매번 마감 기한 전까지 소재 고민하기',
-    afterImg: '/images/experience-summary-after1.jpg',
-    afterCaption: '경험 서랍에서 바로 직무 FIT한 소재 뽑기',
-  },
-  {
-    beforeImg: '/images/experience-summary-before2.jpg',
-    beforeCaption: '여기저기 분산된 경험들',
-    afterImg: '/images/experience-summary-after2.jpg',
-    afterCaption: '경험 마인드맵과 STAR로 논리적인 경험 정리',
-  },
-  {
-    beforeImg: '/images/experience-summary-before3.jpg',
-    beforeCaption: '자기소개서 소재 무한 고민 굴레',
-    afterImg: '/images/experience-summary-after3.jpg',
-    afterCaption: '경험별로 명확하게 연결된 나만의 역량',
-  },
-];
+import BadgedBox from './challenge-result/BadgedBox';
+import {
+  CAREER_START_CONTENT,
+  EXPERIENCE_SUMMARY_CONTENT,
+  PERSONAL_STATEMENT_CONTENT,
+  PORTFOLIO_CONTENT,
+  RESUME_CAREER_START_CONTENT,
+  superTitle,
+} from './challenge-result/constants';
+import ResultImg from './challenge-result/ResultImg';
 
 const {
   PORTFOLIO,
@@ -346,53 +259,6 @@ function ChallengeResult({
         ))}
       </div>
     </section>
-  );
-}
-
-function BadgedBox({
-  badgeContent,
-  badgeColor,
-  className,
-  children,
-  badgeStyle,
-}: {
-  badgeContent: string;
-  badgeColor?: string;
-  className?: string;
-  children?: ReactNode;
-  badgeStyle?: CSSProperties;
-}) {
-  const style = { color: badgeColor, ...badgeStyle };
-
-  return (
-    <Box
-      className={twMerge(
-        'flex w-full min-w-[260px] flex-col overflow-hidden p-0 md:p-0',
-        className,
-      )}
-    >
-      <div
-        className={twMerge(
-          'bg-neutral-75 text-xsmall16 md:text-small20 w-full px-2.5 py-1 text-center font-semibold text-white md:py-2.5',
-        )}
-        style={style}
-      >
-        {badgeContent}
-      </div>
-      {children}
-    </Box>
-  );
-}
-
-function ResultImg({
-  src,
-  alt,
-}: React.DetailedHTMLProps<
-  React.ImgHTMLAttributes<HTMLImageElement>,
-  HTMLImageElement
->) {
-  return (
-    <img className="aspect-[13000/7947] h-full w-full" src={src} alt={alt} />
   );
 }
