+++ apps/web/src/domain/program/challenge/challenge-view/challenge-result/ResultImg.tsx
@@ -1,0 +1,13 @@
+function ResultImg({
+  src,
+  alt,
+}: React.DetailedHTMLProps<
+  React.ImgHTMLAttributes<HTMLImageElement>,
+  HTMLImageElement
+>) {
+  return (
+    <img className="aspect-[13000/7947] h-full w-full" src={src} alt={alt} />
+  );
+}
+
+export default ResultImg;
