+++ apps/web/src/domain/program/challenge/portfolio-view/other-program-recommend/types.ts
@@ -1,0 +1,5 @@
+export type ProgressItemType = {
+  index: number;
+  title: string;
+  subTitle?: string;
+};
